refactor: ProductInventoryService 분리(#104)
